+++ employee-manager/docker-compose.yaml
@@ -1,7 +1,7 @@
 version: '3.7'
 services:
   postgres:
-    image: postgres:10.5
+    image: postgres:latest
     restart: always
     environment:
       - POSTGRES_DB=keploy-test
