+++ employee-manager/src/main/java/com/example/demo/controller/EmployeeController.java
@@ -1,0 +1,65 @@
+package com.example.demo.controller;
+
+import com.example.demo.exception.ResourceNotFoundException;
+import com.example.demo.model.Employee;
+import com.example.demo.repository.EmployeeRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
+import java.time.Instant;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/")
+public class EmployeeController {
+
+    @Autowired
+    private EmployeeRepository employeeRepository;
+
+   //get employees
+   @GetMapping("employees")
+   public List<Employee> getAllEmployee() {
+       return this.employeeRepository.findAll();
+   }
+
+    //get employee by id
+    @GetMapping("employees/{id}")
+    public ResponseEntity<Employee> getEmployeeById(@PathVariable(value = "id") Long employeeId) throws ResourceNotFoundException {
+        Employee employee = employeeRepository.findById(employeeId).orElseThrow(() -> new ResourceNotFoundException("Employee not found for this id :: " + employeeId));
+        return ResponseEntity.ok().body(employee);
+    }
+
+    //save employee
+    @PostMapping("employees")
+    public Employee createEmployee(@RequestBody Employee employee) {
+        employee.setTimestamp(Instant.now().getEpochSecond());
+        return this.employeeRepository.save(employee);
+    }
+
+//    //update employee
+   @PutMapping("employees/{id}")
+   public ResponseEntity<Employee> updateEmployee(@PathVariable(value = "id") Long employeeId, @Valid @RequestBody Employee employeeDetails) throws ResourceNotFoundException {
+
+       Employee employee = employeeRepository.findById(employeeId).orElseThrow(() -> new ResourceNotFoundException("Employee not found for this id :: " + employeeId));
+       employee.setEmail(employeeDetails.getEmail());
+       employee.setFirstName(employeeDetails.getFirstName());
+       employee.setLastName(employeeDetails.getLastName());
+       employee.setTimestamp(Instant.now().getEpochSecond());
+       return ResponseEntity.ok(this.employeeRepository.save(employee));
+   }
+
+//    //delete employee
+   @DeleteMapping("employees/{id}")
+   public Map<String, Boolean> deleteEmployee(@PathVariable(value = "id") Long employeeId) throws ResourceNotFoundException {
+       Employee employee = employeeRepository.findById(employeeId).orElseThrow(() -> new ResourceNotFoundException("Employee not found for this id :: " + employeeId));
+       this.employeeRepository.delete(employee);
+       Map<String, Boolean> response = new HashMap<>();
+       response.put("deleted", Boolean.TRUE);
+
+       return response;
+   }
+}
