+++ employee-manager/src/main/resources/test-runs-success.png
