+++ employee-manager/src/test/java/SamplesJavaApplication_Test.java
@@ -1,0 +1,5 @@
+
+
+public class SamplesJavaApplication_Test {
+
+}
