--- src/main/resources/application.properties
@@ -1,18 +1,0 @@
-spring.datasource.url=jdbc:postgresql://localhost:5439/keploy-test
-spring.datasource.username=keploy-user
-spring.datasource.password=keploy
-#spring.jpa.show-sql=true
-#spring.datasource.driverClassName=io.keploy.ksql.KDriver
-spring.jpa.properties.hibernate.format_sql=true
-## Hibernate Properties
-# The SQL dialect makes Hibernate generate better SQL for the chosen database
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-#Hibernate ddl auto (create, create-drop, validate, update)
-spring.jpa.hibernate.ddl-auto=update
-
-# in order to populate some initial data for testing mark it as true.
-#spring.jpa.defer-datasource-initialization=true
-#spring.sql.init.mode=always
-
-server.port=8080
-logging.level.root=INFO
