--- src/test/e2e/keploy-tests/test-1.yaml
@@ -1,41 +1,0 @@
-version: api.keploy.io/v1beta2
-kind: Http
-name: test-1
-spec:
-    metadata: {}
-    req:
-        method: POST
-        proto_major: 1
-        proto_minor: 1
-        url: /api/employees
-        header:
-            accept: '*/*'
-            content-length: "100"
-            content-type: application/json
-            host: localhost:8080
-            user-agent: curl/7.86.0
-        body: |-
-            {
-                "firstName": "Myke",
-                "lastName": "Tyson",
-                "email": "mt@gmail.com",
-                "timestamp":1
-            }
-        body_type: utf-8
-    resp:
-        status_code: 200
-        header: {}
-        body: '{"id":1,"firstName":"Myke","lastName":"Tyson","email":"mt@gmail.com","timestamp":1677672085}'
-        body_type: utf-8
-        status_message: OK
-        proto_major: 1
-        proto_minor: 1
-    objects:
-        - type: error
-          data: H4sIAAAAAAAA/wEAAP//AAAAAAAAAAA=
-    mocks:
-        - mock-1-0
-    assertions:
-        noise:
-            - body.timestamp
-    created: 1677672085
