--- src/test/e2e/mocks/mock-1.yaml
@@ -1,18 +1,0 @@
-version: api.keploy.io/v1beta1
-kind: SQL
-name: mock-1-0
-spec:
-    metadata:
-        getColumnCount: "5"
-        method: close()
-    type: TABLE
-    table:
-        cols:
-            - name: id
-              type: Long
-              precision: 0
-              scale: 0
-        rows:
-            - '[`1`]'
-    int: 1
-    error: []
