--- src/test/e2e/mocks/mock-2.yaml
@@ -1,30 +1,0 @@
-version: api.keploy.io/v1beta1
-kind: SQL
-name: mock-2-0
-spec:
-    metadata:
-        SQL-Query: select employee0_.id as id1_0_0_, employee0_.email as email2_0_0_, employee0_.first_name as first_na3_0_0_, employee0_.last_name as last_nam4_0_0_, employee0_.timestamp as timestam5_0_0_ from employees employee0_ where employee0_.id=?
-        method: next()
-    type: TABLE
-    table:
-        cols:
-            - name: email2_0_0_
-              type: String
-              precision: 0
-              scale: 0
-            - name: first_na3_0_0_
-              type: String
-              precision: 0
-              scale: 0
-            - name: last_nam4_0_0_
-              type: String
-              precision: 0
-              scale: 0
-            - name: timestam5_0_0_
-              type: Long
-              precision: 0
-              scale: 0
-        rows:
-            - '[`mt@gmail.com`|`Myke`|`Tyson`|`1677672085`]'
-    int: 0
-    error: []
